requirements.txt
actionlib==1.11.13
angles==1.9.11
backports.functools-lru-cache==1.5
bondpy==1.8.3
camera-calibration-parsers==1.11.13
catkin==0.7.14
catkin-pkg==0.4.12
certifi==2019.3.9
chainer==6.0.0
chainercv==0.13.0
chardet==3.0.4
cloudpickle==1.2.1
cv-bridge==1.12.8
cycler==0.10.0
Cython==0.29.11
decorator==4.4.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.14
dynamic-reconfigure==1.5.50
fcn==6.4.16
filelock==3.0.12
future==0.17.1
gdown==3.8.1
gencpp==0.6.0
geneus==2.2.6
genjava==0.3.4
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
idna==2.8
image-geometry==1.12.8
interactive-markers==1.11.4
joblib==0.13.2
kiwisolver==1.1.0
laser-geometry==1.6.4
matplotlib==2.2.4
message-filters==1.12.14
natsort==6.0.0
networkx==2.2
nose==1.3.7
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
Pillow-PIL==0.1.dev0
pluginlib==1.11.3
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-qt-binding==0.3.4
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.22.0
resource-retriever==1.12.4
rosbag==1.12.14
rosboost-cfg==1.14.4
rosclean==1.14.4
roscreate==1.14.4
rosgraph==1.12.14
rosjava-build-tools==0.3.3
roslaunch==1.12.14
roslib==1.14.4
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.4
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.1.9
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.4
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.17
scikit-image==0.14.3
scikit-learn==0.20.3
scipy==1.2.2
sensor-msgs==1.12.7
setproctitle==1.1.10
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
subprocess32==3.5.4
tensorboardX==1.7
tf==1.11.9
tf-conversions==1.11.9
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.1
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.25.3
xacro==1.11.3
